requirements.txt
langchain==0.3.27
langchain-community==0.3.27
langchain-google-genai==2.1.10
langchain-openai==0.3.32
langchain-postgres==0.0.15
langchain-text-splitters==0.3.9
openai==1.107.2
psycopg[binary]==3.2.9
pypdf==5.9.0
python-dotenv==1.0.1
